Extract /integrations code from landing-page into separate bundle.
