+++ templates/zerver/integrations/index.html
@@ -6,7 +6,7 @@
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 {% stylesheet 'portico' %}
 {% stylesheet 'landing-page' %}
-{{ render_bundle('landing-page') }}
+{{ render_bundle('integrations') }}
 
 {% endblock %}
